requirements.txt
requests==2.31.0
websocket-client==1.6.2
pandas==2.0.3
numpy==1.23.5
tqdm==4.66.1
mysql.connector==2.2.9
python-dotenv==1.0.0
pytest==7.4.0
tensorflow==2.12.0
sqlalchemy==2.0.19
fastapi==0.101.1
uvicorn==0.23.2
Jinja2==3.1.2
plotly==5.16.1
python-multipart==0.0.6